refactor(video): park the v2 composition under src/legacy as WalkthroughV2
The v3 cut takes over the primary name. The v2 composition and its
primitives, MusicBed included, stay read-only until no v2 artifacts
remain; the legacy renderer path targets WalkthroughV2.
